+++ client/src/scripts/components/profile/Chart.jsx
@@ -1,4 +1,11 @@
 // Chart.js
+// Reflux
+var Reflux = require('reflux');
+// Router
+var Router = require('react-router');
+// Actions
+var actions = require('../../actions/actions');
+
 var D3Chart = require('./D3Chart');
 
 var Chart = React.createClass({
