+++ client/src/scripts/components/profile/Dashboard.jsx
@@ -1,3 +1,10 @@
+// Reflux
+var Reflux = require('reflux');
+// Router
+var Router = require('react-router');
+// Actions
+var actions = require('../../actions/actions');
+
 // Components
 var Chart = require('./Chart');
 var Tracker = require('./Tracker');
