+++ client/src/scripts/components/profile/Footer.jsx
@@ -1,3 +1,10 @@
+// Reflux
+var Reflux = require('reflux');
+// Router
+var Router = require('react-router');
+// Actions
+var actions = require('../../actions/actions');
+
 var Footer = React.createClass({
   render: function() {
     return (
