+++ client/src/scripts/components/profile/radial.jsx
@@ -1,3 +1,10 @@
+// Reflux
+var Reflux = require('reflux');
+// Router
+var Router = require('react-router');
+// Actions
+var actions = require('../../actions/actions');
+
 var d3 = require('d3');
 
 var Radial = {};
