--- client/src/scripts/components/profile/style.css
@@ -1,61 +1,0 @@
-/**
- Copyright (c) 2014 BrightPoint Consulting, Inc.
-
- Permission is hereby granted, free of charge, to any person
- obtaining a copy of this software and associated documentation
- files (the "Software"), to deal in the Software without
- restriction, including without limitation the rights to use,
- copy, modify, merge, publish, distribute, sublicense, and/or sell
- copies of the Software, and to permit persons to whom the
- Software is furnished to do so, subject to the following
- conditions:
-
- The above copyright notice and this permission notice shall be
- included in all copies or substantial portions of the Software.
-
- THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- OTHER DEALINGS IN THE SOFTWARE.
- */
-
-.background {
-    fill: #FFFFFF;
-    fill-opacity: 0.01;
-}
-
-.component {
-    fill: #e1e1e1;
-}
-
-.component .label {
-    font-family: Myriad, "Helvetic Neue", Helvetica, Arial;
-    text-anchor: middle;
-    fill: #0000FF;
-}
-
-.arc {
-    stroke-weight:1;
-    fill: #4e8fff;
-}
-
-
-.arc2 {
-    stroke-weight:1;
-    fill: #3660b0;
-}
-
-
-.label {
-    font-family:  Myriad, "Helvetic Neue", Helvetica, Arial;
-    text-anchor: middle;
-}
-
-.radial-svg {
-    display: block;
-    margin: 0 auto;
-}
