+++ client/src/scripts/views/connect.jsx
@@ -12,9 +12,9 @@ var actions = require('../actions/actions');
 var loginStore = require('../stores/loginStore');
 // Components
 var Spinner = require('../components/spinner');
-var Jawbone = require('../components/devices/jawbone');
-var Fitbit = require('../components/devices/fitbit');
-var Other = require('../components/devices/other');
+var Jawbone = require('../components/Devices/Jawbone');
+var Fitbit = require('../components/Devices/Fitbit');
+var Other = require('../components/Devices/Other');
 /**
 @description This is where users will be able to set up their application.
 Keep an eye out for <Fitbit /> and <Jawbone />
